+++ test/integration/nft.test.ts
@@ -622,6 +622,16 @@ describe('E2E integration tests', () => {
     expect(typeof response.contractAddresses[0]).toEqual('string');
   });
 
+  it('reportSpam()', async () => {
+    const response = await alchemy.nft.reportSpam(contractAddress);
+    expect(typeof response).toEqual('undefined');
+  });
+
+  it('isAirdropNft()', async () => {
+    const response = await alchemy.nft.isAirdropNft(contractAddress, '0');
+    expect(typeof response.isAirdrop).toEqual('boolean');
+  });
+
   it('getFloorPrice()', async () => {
     const response = await alchemy.nft.getFloorPrice(contractAddress);
     expect(response.openSea).toBeDefined();